do not add disabled and hidden deps to a return set that does not define the one used to build the models
The DataSetFilterUtil does not constrain the dataSet used when building world models,
so we do not need to care about adding DisabledAndHidden primitives to the returned
set at this point.

Since the filter utility _must_ only reply objects that match one of the layerMatcher
functions, ..
- getNodeMatcher()
- getWayMatcher()
- getRelationMatcher()

.. the previously added DisabledAndHidden primitives should have been passed through
these matchers before adding them to the result set anyway.

In other words: A disabled and hidden node of an enabled and visible way (that passes
the matcher) does not need to be added to the return set at this point for the way to
build at a later stage.  The same applies for disabled and hidden children of
relations.