enable two sided lighting on a per-model basis, option to choose how fences are drawn
fence drawing now works both ways:
- with a single face and double sided lighting
  for the Wall, Fence and FenceRelation models
- with two faces and face culling

there is an expert option in josm preferences
to choose which one is used:

kendzi3d.models.preferTwoSidedLightingOverFaceCulling

This is enabled by default, because it is more
performant.  If it does not work for some reason,
user has the choice to set this option to "false"
and the face culling method is re-enabled.

Forcing two sided lighting has been renamed for
clarity and made an expert option (if josm expert
mode is disabled, menu entry does not show in 3D
menu).

The camera layer fix suggested in response to the
first attempt has been merged / included.

restored to original state:
- model/ground/GroundDrawer.java
- model/ground/SelectableGround.java
- model/ground/StyledTitleGroundDrawer.java

(ForceTwoSidedLighting option can now be applied
to models derived from AbstractModel.java only,
so we can keep the ground rendering with a one
sided lighting calculation.

The previous need for this existed, because two
sided lighting was enabled unconditionally for
this action in the material setup routines.

Two sided lighting is not set up in draw()
method, just as cull faces is.)